+++ json_serializable/lib/src/type_helpers/enum_helper.dart
@@ -4,7 +4,6 @@
 
 import 'package:analyzer/dart/element/type.dart';
 
-import '../constants.dart' as consts;
 import '../type_helper.dart';
 import '../utils.dart';
 
@@ -31,18 +30,9 @@ class EnumHelper extends TypeHelper {
       return null;
     }
 
-    var wrappedExpression =
-        simpleExpression.hasMatch(expression) ? expression : '{$expression}';
-
-    var closureArg = consts.closureArg;
-    if (closureArg == wrappedExpression) {
-      closureArg = '${closureArg}2';
-    }
-
-    return commonNullPrefix(
-        context.nullable,
-        expression,
-        '$targetType.values.singleWhere(($closureArg) => $closureArg.toString()'
-        " == '$targetType.\$$wrappedExpression')");
+    var functionName =
+        context.nullable ? r'$enumDecodeNullable' : r'$enumDecode';
+    return "$functionName('$targetType', $targetType.values, "
+        '$expression as String)';
   }
 }