+++ json_serializable/test/default_value/default_value.checked.dart
@@ -5,19 +5,27 @@
 // GENERATED CODE - DO NOT MODIFY BY HAND
 
 // **************************************************************************
-// Generator: _NonNullableGenerator
+// Generator: _CheckedGenerator
 // **************************************************************************
 
 // ignore_for_file: annotate_overrides
 
 import 'package:json_annotation/json_annotation.dart';
 
-part 'default_value.non_nullable.g.dart';
+import 'default_value_interface.dart' as dvi hide Greek;
+import 'default_value_interface.dart' show Greek;
+
+part 'default_value.checked.g.dart';
 
 const _intValue = 42;
 
-@JsonSerializable(nullable: false)
-class DefaultValue extends Object with _$DefaultValueSerializerMixin {
+dvi.DefaultValue fromJson(Map<String, dynamic> json) =>
+    _$DefaultValueFromJson(json);
+
+@JsonSerializable()
+class DefaultValue extends Object
+    with _$DefaultValueSerializerMixin
+    implements dvi.DefaultValue {
   @JsonKey(defaultValue: true)
   bool fieldBool;
 
@@ -46,6 +54,9 @@ class DefaultValue extends Object with _$DefaultValueSerializerMixin {
     'root': ['child']
   })
   Map<String, List<String>> fieldMapListString;
+
+  @JsonKey(defaultValue: Greek.beta)
+  Greek fieldEnum;
 
   DefaultValue();