+++ json_serializable/test/json_serializable_test.dart
@@ -613,24 +613,50 @@ abstract class _$SubTypeSerializerMixin {
     expect(output, expected);
   });
 
-  group('default values', () {
-    test('fail with symbols', () {
-      expectThrows(
-          'DefaultWithSymbol',
-          'Error with `@JsonKey` on `field`. '
-          'Values of type `Symbol` are not supported for `defaultValue`.');
+  if (!generator.useWrappers) {
+    group('default values fail with', () {
+      test('symbols', () {
+        expectThrows(
+            'DefaultWithSymbol',
+            'Error with `@JsonKey` on `field`. '
+            '`defaultValue` is `Symbol`, it must be a literal.');
+      });
+      test('functions', () {
+        expectThrows(
+            'DefaultWithFunction',
+            'Error with `@JsonKey` on `field`. '
+            '`defaultValue` is `Function`, it must be a literal.');
+      });
+      test('type', () {
+        expectThrows(
+            'DefaultWithType',
+            'Error with `@JsonKey` on `field`. '
+            '`defaultValue` is `Type`, it must be a literal.');
+      });
+      test('const object', () {
+        expectThrows(
+            'DefaultWithConstObject',
+            'Error with `@JsonKey` on `field`. '
+            '`defaultValue` is `Duration`, it must be a literal.');
+      });
+      test('enum value', () {
+        expectThrows(
+            'DefaultWithNestedEnum',
+            'Error with `@JsonKey` on `field`. '
+            '`defaultValue` is `List > Enum`, it must be a literal.');
+      });
+      test('non-nullable field', () {
+        expectThrows(
+            'DefaultWithNonNullableField',
+            'Error with `@JsonKey` on `field`. '
+            'Cannot use `defaultValue` on a field with `nullable` false.');
+      });
+      test('non-nullable class', () {
+        expectThrows(
+            'DefaultWithNonNullableClass',
+            'Error with `@JsonKey` on `field`. '
+            'Cannot use `defaultValue` on a field with `nullable` false.');
+      });
     });
-    test('fail with functions', () {
-      expectThrows(
-          'DefaultWithFunction',
-          'Error with `@JsonKey` on `field`. '
-          'The provided `defaultValue` is not a literal: () → int (_function)');
-    });
-    test('fail with type', () {
-      expectThrows(
-          'DefaultWithType',
-          'Error with `@JsonKey` on `field`. '
-          'Values of type `Type` are not supported for `defaultValue`.');
-    });
-  });
+  }
 }